requirements.txt
scikit_learn == 1.2.2
pandas == 2.1.4
numpy == 1.26.3
mlflow == 2.6.0
pytorch == 2.1.2